// apps/web/src/entities/product/model/product.ts
export interface ProductBaseInfo {
    /** 판매자관리코드 */
    productManagementCd: string;
    /** 원산지(기타 직접입력, 수입사등 표기) */
    placeOriginEtcLabel: string;
    /** 상품유형 */
    productClassType: ProductType;
    /** 상품정보고시 (JSON.stringify) */
    dutyInfo: string;
    /** 쿠폰 적용 가능 여부 (Y : 쿠폰 적용 가능, N : 쿠폰 적용 불가) */
    couponUseYn: OptionYnType;
    /** 상품설명HTML 상단 */
    contentHeader: string;
    /** 옵션이미지상세보기여부 */
    optionImageViewable: boolean;
    /** 상품설명HTML 본문 */
    content: string;
    /** 상품명 */
    productName: string;
    /** 미성년자 구매 가능 여부 (Y : 미성년자 구매 가능, N : 미성년자 구매 불가) */
    minorPurchaseYn: OptionYnType;
    /** 판매자 특이사항/고객안내사항 */
    deliveryCustomerInfo: string;
    /** 상품 이미지URL 타입 */
    imageUrlInfo: ImageUrlType[];
    /** 적립금 사용 한도율 (nullable) */
    accumulationUseLimitInfo: Nullable<{
        /** 적립금 사용 금액 단위, (AMOUNT: One, PERCENT: %) */
        unitType: string;
        /** 적립금 사용 양 */
        limitValue: number;
    }>;
    /** 상품번호 */
    productNo: number;
    /** 상품항목추가 관리 정보 */
    customPropertise: {
        /** 상품 항목 값 번호 */
        propValueNo?: number;
        /** 항목 복수선택여부 (Y: 복수개 선택가능, N: 1개만 선택가능) */
        multipleSelectionYn?: OptionYnType;
        /** 추가 항목 타입, Enum: [ STRING: 문자열, COLOR: 컬러 ] */
        propType?: CustomPropertiesPropType;
        /** 상품 항목 값 */
        propValue?: string;
        /** 상품 항목명 */
        propName?: string;
        /** 상품 항목명 번호 */
        propNo?: number;
    }[];
    /** 재입고 알림 사용 가능 여부 - true(사용가능), false(불가능) */
    usableRestockNoti: boolean;
    /** 유효일자 */
    expirationYmdt: string;
    /** 상품 등록일 */
    registerYmdt: string;
    /** 상품설명HTML 하단 */
    contentFooter: string;
    /** 인증 타입 (TARGET: 인증대상, NOT_TARGET: 인증대상아님, DETAIL_PAGE: 상세페이지 별도표기) */
    certificationType: CertificationType;
    /** 제조일자 */
    manufactureYmdt: string;
    /** 적립금 사용 여부 (Y:사용가능 , N:사용불가능) */
    accumulationUseYn: OptionYnType;
    /** 스티커 라벨(배열) */
    stickerLabels: string[];
    /** URL로만 접근 여부 (Y : URL로만 접근 전용, N : 일반) */
    urlDirectDisplayYn: OptionYnType;
    certifications: CertificationType[];
    /** 판매종료일시 */
    saleEndYmdt: string;
    /** 홍보문구 */
    promotionText: string;
    /** 판매기간 타입 */
    salePeriodType: SalePeriodType;
    /** 스티커 정보 */
    stickerInfos: StickerInfo[];
    /** 결제수단 설정 정보 */
    paymentMeans: PayType;
    /** HS CODE */
    hsCode: string;
    /** 상품군 (DELIVERY: Delivery Group, SERVICE: Service Product Group) */
    productGroup: ProductGroupType;
    /** 판매시작일시 */
    saleStartYmdt: string;
    /** 상품 이미지URL */
    imageUrls: string[];
    /** 상품등록유형 */
    mappingType: MappingType;
    /** 원산지 */
    placeOriginLabel: string;
    /** 영문 상품명 */
    productNameEn: string;
    /** 구매 안내 */
    purchaseGuide: string;
}

export interface ProductShippingInfo {
    /** 배송 설정 */
    shippingConfig: {
        /** 배송 템플릿 번호 */
        templateNo: number;
        /** 배송지 파트너 번호 */
        shippingAreaPartnerNo: number;
        /** 출고 유형(배송 구분) */
        shippingAreaType: ShippingAreaType;
        /** 해외 배송 여부 (true: 해외 배송, false:국내 배송) */
        internationalShippingAvailable: boolean;
        /** 묶음배송 가능여부 (true: 묶음 배송 가능, false: 묶음 배송 불가능) */
        combinable: boolean;
    };
    /** 배송 가능 여부 */
    shippingAvailable: boolean;
}

export interface ProductStock {
    /** 구매 수량 (재고 미노출의 경우 -999 재고 미노출 설정일때 실재고가 없는 경우, 0으로 표기) */
    saleCnt: number;
    /** 대표 남은 수량(옵션의 추가금이 0인 재고의 합/재고 미노출의 경우 -999 재고 미노출 설정일때 실재고가 없는 경우, 0으로 표기) */
    mainStockCnt: number;
    /** 남은 수량 (재고 미노출의 경우 -999 재고 미노출 설정일때 실재고가 없는 경우, 0으로 표기) */
    stockCnt: number;
}

export interface ProductDeliveryDate {
    period: {
        /** 시작일 */
        startYmdt: string;
        /** 종료일 */
        endYmdt: string;
    };
    /** 요일 */
    daysOfWeek: DaysOfWeekType;
    /** 주문일 기준 */
    daysAfterPurchase: Nullable<number>;
}

export interface Brand {
    /** 브랜드명 유형 */
    nameType: 'NAME_KO' | 'NAME_EN' | 'NONE';
    /** 브랜드명 */
    name: string;
    nameKo: string;
    /** 영문 브랜드명 */
    nameEn: string;
    /** 브랜드번호 */
    brandNo: number;
    /** 브랜드 로고 이미지 URL */
    logoImageUrl: string;
}

export interface ProductLimitations {
    /** 주문환불불가 목록 */
    nonRefundTypes: Nullable<(boolean | string | number)[]>;
    /** 최소구매수량 */
    minBuyCnt: number;
    /** 1회최대구매수량 */
    maxBuyTimeCnt: number;
    /** 최대구매기간(일) */
    maxBuyDays: number;
    /** 최대구매기간(수량) */
    maxBuyPeriodCnt: number;
    /** 1인최대구매수량 */
    maxBuyPersonCnt: number;
    /** 환불가능여부 */
    refundable: boolean;
    /** 네이버페이 결제 가능여부 */
    naverPayHandling: boolean;
    /** 비회원구매여부(true : 가입한 회원만 구매 가능) */
    memberOnly: boolean;
    /** 장바구니 가능 여부 */
    canAddToCart: boolean;
}

/** 상품 카운트 정보 */
// TO CHECK: optional 타입인지 확인
export interface ProductCounter {
    /** 내 상품문의 카운트(accessToken 없을 시, 0) */
    myInquiryCnt?: number;
    /** 상품문의 카운트 */
    inquiryCnt?: number;
    /** 좋아요 수 */
    likeCnt?: number;
    /** 상품평 카운트 */
    reviewCnt?: number;
}

export interface Partner {
    /** 파트너번호 */
    partnerNo: number;
    /** 대표자명 */
    ownerName: string;
    /** FAX번호 */
    faxNo: string;
    /** 판매자명 */
    partnerName: string;
    /** 사업장 주소 */
    officeAddressLabel: string;
    /** 상호명 */
    companyName: string;
    /** 사업자번호 */
    businessRegistrationNo: string;
    /** 통신판매신고번호 */
    onlineMarketingBusinessDeclarationNo: string;
    /** 대표 이메일 */
    email: string;
    /** 대표번호 */
    phoneNo: string;
}

export interface ProductDetailResponse {
    /** 렌탈 정보 (옵션이 없는 상품의 경우 조회, 옵션이 있는 상품의 경우 옵션 조회 API(/products/{productNo}/options) 에서 렌탈 정보 조회 가능) */
    rentalInfos: RentalInfo[];
    /** 예약판매정보 */
    reservationData: ReservationData;
    /** 상품 기본 정보 */
    baseInfo: ProductBaseInfo;
    /** 그룹관리코드 노출명 */
    groupManagementCodeName: Nullable<string>;
    /** 배송 관련 정보 */
    shippingInfo: ProductShippingInfo;
    /** 배송 안내 */
    deliveryGuide: Nullable<string>;
    /** 그룹관리코드 */
    groupManagementCode: Nullable<string>;
    /** 관련 상품 번호 */
    relatedProductNos: number[];
    /** 좋아요 여부(accessToken 없을 시 false) */
    liked: boolean;
    /** 사입 위탁 구분 값 (PURCHASE: purchase, CONSIGNMENT: Consignment) */
    saleMethodType: SaleMethodType;
    /** 교환 안내 */
    exchangeGuide: Nullable<string>;
    /** 환불 안내 */
    refundGuide: Nullable<string>;
    /** 가격정보 */
    price: Price;
    /** 리뷰 작성 가능 여부 */
    reviewAvailable: boolean;
    /** 카테고리 목록 */
    categories: DisplayCategory[];
    /** 재고정보 */
    stock: ProductStock;
    /** 기간 */
    deliveryDate: ProductDeliveryDate;
    /** 브랜드 정보 */
    brand: Brand;
    /** 구매제한 */
    limitations: ProductLimitations;
    /** 상품평 평균점 */
    reviewRate: number;
    /** 메인 베스트 상품 여부 */
    mainBestProductYn: boolean;
    /** 정기 결제 정보 (해당 값이 null로 오느냐에 따라서 정기결제상품인지 아닌지 여부를 판단) */
    regularDelivery: Nullable<{
        /** 정기 결제 즉시 할인 정보 (nullable) */
        discount: Nullable<{
            /** 즉시 할인 단위 (AMOUNT: One, PERCENT: %) */
            type: AccumulationUnitType;
            /** 즉시 할인 금액/율 */
            value: number;
        }>;
    }>;
    /** 상품 카운트 정보 */
    counter: ProductCounter;
    /** 파트너사 공지 */
    partnerNotice: {
        /** 제목 */
        title: string;
        /** 내용 */
        content: string;
    };
    /** AS 안내 */
    afterServiceGuide: Nullable<string>;
    /** 배송정보 */
    deliveryFee: DeliveryFee;
    /** 판매자 정보 */
    partner: Partner;
    /** 주류 통신판매 명령 위임고시 */
    liquorDelegationGuide: Nullable<string>;
    /** 재고 노출 여부 (false:재고 미노출 / true:재고 노출) */
    displayableStock: boolean;
    /** 상품 상태 */
    status: Status;
}
// ...
export interface Category {
    /** 세부 카테고리 정보 */
    categories: {
        /** 뎁스 */
        depth: number;
        /** 카테고리 번호 */
        categoryNo: number;
        /** 카테고리 명 */
        label: string;
    }[];
    /** 전체 카테고리 명 */
    fullCategoryLabel: string;
}
export interface DisplayCategory extends Category {
    /** 대표 전시 카테고리 여부 */
    representativeYn: 'Y' | 'N';
}